+++ video/src/main/java/com/github/makewheels/video2022/video/like/LikeType.java
@@ -1,0 +1,6 @@
+package com.github.makewheels.video2022.video.like;
+
+public interface LikeType {
+    String LIKE = "LIKE";
+    String DISLIKE = "DISLIKE";
+}